Use ListItem instead of Proposal.
#166
