+++ src/hook/useReferendums.ts
@@ -1,0 +1,8 @@
+import {useApi} from 'context/ChainApiContext';
+import {useQuery} from 'react-query';
+
+export function useReferendums() {
+  const {api} = useApi();
+
+  return useQuery('referendums', () => api?.derive.democracy.referendums());
+}
