+++ src/screen/ReferendaScreen.tsx
@@ -1,16 +1,15 @@
-import * as React from 'react';
-import {Divider, Layout, Text} from '@ui-kitten/components';
-import SafeView, {noTopEdges} from 'presentational/SafeView';
-import {FlatList, StyleSheet} from 'react-native';
-import globalStyles, {standardPadding} from 'src/styles';
-import {useApi} from 'context/ChainApiContext';
-import {useQuery} from 'react-query';
-import {EmptyView} from 'presentational/EmptyView';
 import type {DeriveReferendumExt} from '@polkadot/api-derive/types';
-import {Proposal} from 'presentational/Proposal';
-import {useBestNumber} from 'src/hook/useVotingStatus';
 import {BN_ONE} from '@polkadot/util';
+import {Card, Divider, Layout, ListItem, Text} from '@ui-kitten/components';
+import {EmptyView} from 'presentational/EmptyView';
+import {ProposalInfo} from 'presentational/ProposalInfo';
+import SafeView, {noTopEdges} from 'presentational/SafeView';
+import * as React from 'react';
+import {FlatList, StyleSheet} from 'react-native';
 import {useBlockTime} from 'src/hook/useBlockTime';
+import {useReferendums} from 'src/hook/useReferendums';
+import {useBestNumber} from 'src/hook/useVotingStatus';
+import globalStyles, {standardPadding} from 'src/styles';
 
 export function ReferendaScreen() {
   const {data, isLoading, refetch} = useReferendums();
@@ -37,30 +36,32 @@ export function ReferendaScreen() {
 
 const styles = StyleSheet.create({flatList: {padding: standardPadding * 2}});
 
-function useReferendums() {
-  const {api} = useApi();
-
-  return useQuery('referendums', () => api?.derive.democracy.referendums());
-}
-
 function Referenda({item}: {item: DeriveReferendumExt}) {
   const {image: {proposal} = {proposal: undefined}} = item;
   const bestNumber = useBestNumber();
 
   const remainBlock = bestNumber ? item.status.end.sub(bestNumber).isub(BN_ONE) : undefined;
   const {timeStringParts} = useBlockTime(remainBlock);
+  if (!proposal) {
+    return <Text>{item.imageHash.toString()}</Text>;
+  }
 
-  return proposal ? (
-    <Proposal
-      proposal={proposal}
-      accessoryLeft={() => <Text category={'h4'}>{item.index.toString()}</Text>}
-      accessoryRight={() => (
-        <Text category={'c1'} adjustsFontSizeToFit={true} numberOfLines={1}>
-          {timeStringParts.slice(0, 2).join(' ')}
-        </Text>
-      )}
-    />
-  ) : (
-    <Text>{item.imageHash.toString()}</Text>
+  const {method, section} = proposal?.registry.findMetaCall(proposal.callIndex);
+
+  return (
+    <Card style={referendumStyle.container}>
+      <ListItem
+        title={`${method}.${section}`}
+        accessoryLeft={() => <Text category={'h4'}>{item.index.toString()}</Text>}
+        accessoryRight={() => (
+          <Text category={'c1'} adjustsFontSizeToFit={true} numberOfLines={1}>
+            {timeStringParts.slice(0, 2).join(' ')}
+          </Text>
+        )}
+      />
+      <ProposalInfo proposal={proposal} />
+    </Card>
   );
 }
+
+const referendumStyle = StyleSheet.create({container: {marginBottom: standardPadding}});
